If we have it, use history.replaceState to live-update the URL.
Update the URL in-place as the user searches or selects filters, so that
users can just copy the address from the address bar and get the current
page they're looking at.

Keep the button and box for browsers that don't have replaceState yet,
but we could also plausibly just drop it.