Filter for minimum and maximum values.
Based on https://github.com/powdahound/ec2instances.info/issues/132
but had to tweak it a bit. I know it is really ugly but works.

I think we need a real frontend dev, which would be able to apply
properly https://datatables.net/examples/plug-ins/range_filtering.html
But this would rather require major redesinging of the html and json.