Switch from Fabric to Invoke
https://www.fabfile.org/upgrading.html#sidegrading-to-invoke

This reduces the number of dependencies needed for the project and also removes
the first stumbling block for running the project with Python 3.

Invoke makes use of a `Context` object which contains the context where the
tasks are run inside of. It is supplied to the tasks as the first positional
argument so I had to add a `c` argument to all the tasks to make this work.

Invoke doesn't have an `abort` function like Fabric has, so I have instead
simplified it to just print the abort message and then exit with a status code
of 1.