Use the instance type data to populate the architecture list
AWS currently has support for both the x86 and ARM architectures,
so the old assumption that everything was x86_64 unless there was
a 32-bit pricing option doesn't apply.  Fortunately, AWS provides
the list of supported architectures as part of the DescribeInstanceTypes
call, so we can will support any new architectures that they come
out with.