Fix Spot instance pricing for instances such as p3.16xlarge
This was caused by the use of `insort`, which legitimately considers the
Spot data as strings, as per instances.json, essentially giving us string
alphabetic sorting instead of numeric sorting.

The sorting is now done after each insert, costing us a few more cycles,
but this way we can force the sorting to be done by numerical value.

Fixes #578.